mattp - #35 move files around to respective pages
