+++ src/components/common/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { Navigate, Outlet } from "react-router-dom";
-import { AuthContext } from "../contexts/AuthContext";
+import { AuthContext } from "../../contexts/AuthContext";
 
 export default function ProtectedRoute() {
   const authContext = useContext(AuthContext);
