+++ src/pages/LoginPage.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { useAuth } from "../../hooks/useAuth";
+import { useAuth } from "../hooks/useAuth";
 
 const LoginPage = () => {
   const [username, setUsername] = useState("");
